backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pandas==2.2.2
numpy==1.26.4
duckdb==1.1.0
pyod==2.0.0
scikit-learn==1.5.2
google-generativeai==0.8.3
groq>=0.13.0
weasyprint==62.3
python-dotenv==1.0.1
python-multipart==0.0.12

openai==1.55.0
httpx>=0.23.0,<0.28.0
python-docx==1.1.0
